requirements.txt
aiogram>=3.0,<4.0
requests>=2.31,<3
python-dotenv>=1.0,<2
